fix(uid): remove uid$ from app reducer and facade
The uid$ was causing issues on logout since PR #15. It was causing 6+ errors when a user logs out
and the state is trying to locate an undefined uid$. It is removed from app.facade and app.reducer.
The only other additional place that it was used was in history.service in which it was replaced by
the authFacade.user in which contains the UID needed for that logic.